LayoutTests/platform/glib/fast/mediastream/RTCPeerConnection-remotely-assigned-transceiver-mid-expected.txt

Test the case where an RTCRtpTransceiver gets a remotely assigned mid and also unmute the associated source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pcA.addTrack(stream.getAudioTracks()[0], stream)
A's transceiver is not yet associated with a media description and should have a null mid
PASS pcA.getTransceivers()[0].mid is null
A: local offer set (mid gets defined)
midAssignedByA = pcA.getTransceivers()[0].mid
FAIL midAssignedByA !== null should be true. Was false.
pcB.addTrack(stream.getAudioTracks()[0], stream)
B's transceiver is not yet associated with a media description and should have a null mid
PASS pcB.getTransceivers()[0].mid is null
B: remote offer set (mid gets defined)
B's transceiver should get its mid from the remote side (A)
PASS pcB.getTransceivers()[0].mid is midAssignedByA
PASS Offer/answer dialog completed
PASS B: got remote track event
PASS event.track is an instance of MediaStreamTrack
PASS A: got remote track event
PASS B: remote track unmute event
PASS successfullyParsed is true

TEST COMPLETE
